chore(crypto): Unify testing infrastructure for `raw_signature` (#848)
