+++ internal/crypto/src/tests/raw_signature/mod.rs
@@ -11,4 +11,7 @@
 // specific language governing permissions and limitations under
 // each license.
 
-mod validator;
+mod async_signers;
+mod async_validators;
+mod signers;
+mod validators;
